+++ lib/form/index.ts
@@ -1,4 +1,3 @@
 export * from "./context";
 export * from "./form";
-export * from "./use-formlike";
 export * from "./types";
